refactor: 패키지 구조 변경
- movie패키지 추가
