+++ app/src/androidTest/java/woowacourse/movie/ui/reservation/MovieReservationActivityTest.kt
@@ -19,9 +19,9 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
-import woowacourse.movie.model.MovieContent
-import woowacourse.movie.model.MovieDate
 import woowacourse.movie.model.data.MovieContentsImpl
+import woowacourse.movie.model.movie.MovieContent
+import woowacourse.movie.model.movie.MovieDate
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
