+++ app/src/main/java/woowacourse/movie/model/movie/MovieContent.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.movie
 
 data class MovieContent(
     val imageId: Int,
