+++ app/src/main/java/woowacourse/movie/model/movie/MovieDate.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.movie
 
 data class MovieDate(
     val year: Int,
