+++ app/src/main/java/woowacourse/movie/model/movie/Ticket.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.movie
 
 class Ticket(
     val reservationCount: ReservationCount,
