+++ app/src/test/java/woowacourse/movie/model/ReservationCountTest.kt
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import woowacourse.movie.model.movie.ReservationCount
 
 class ReservationCountTest {
     @ParameterizedTest
